+++ rulesets/tests/test_enclosure.rb
@@ -1189,9 +1189,9 @@ class ERIEnclosureTest < MiniTest::Test
   def _check_floors(hpxml, area:, rvalue:)
     area_values = []
     rvalue_x_area_values = [] # Area-weighted
-    hpxml.frame_floors.each do |frame_floor|
-      area_values << frame_floor.area
-      rvalue_x_area_values << frame_floor.insulation_assembly_r_value * frame_floor.area
+    hpxml.floors.each do |floor|
+      area_values << floor.area
+      rvalue_x_area_values << floor.insulation_assembly_r_value * floor.area
     end
 
     assert_in_epsilon(area, area_values.inject(:+), 0.01)